

\name{science1.item.characteristics}

\docType{data}

\alias{science1.item.characteristics}

\title{Science achievement test data - Item characteristics}

\description{
This data frame contains item characteristics for usage with \code{\link{automateModels}} and data set \code{\link{science1}}
}

%\usage{ }

\format{'data.frame':   185 obs. of  3 variables}

%\source{}

%\references{
%}

\keyword{datasets}






